Address Worth Checking Out tooltip CI feedback
