+++ components/brain/left-sidebar/web/WebUnifiedWavesListWaves.tsx
@@ -1,12 +1,11 @@
 "use client";
 
 import PrimaryButton from "@/components/utils/button/PrimaryButton";
-import HoverCard from "@/components/utils/tooltip/HoverCard";
 import { useMyStream } from "@/contexts/wave/MyStreamContext";
 import useCreateModalState from "@/hooks/useCreateModalState";
 import useIsTouchDevice from "@/hooks/useIsTouchDevice";
 import { usePrefetchWaveData } from "@/hooks/usePrefetchWaveData";
-import { faInfoCircle, faPlus } from "@fortawesome/free-solid-svg-icons";
+import { faPlus } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import Link from "next/link";
 import React, { useCallback, useMemo, useRef } from "react";
@@ -20,6 +19,7 @@ import {
   getHighlyRatedPreviewWaves,
   HighlyRatedWavesToggle,
 } from "../waves/HighlyRatedWavesToggle";
+import { SidebarCategoryLabel } from "../waves/SidebarCategoryLabel";
 import { SidebarWaveRowsSection } from "../waves/SidebarWaveRowsSection";
 import SectionHeader from "../waves/SectionHeader";
 import WavesFilterToggle from "../waves/WavesFilterToggle";
@@ -58,16 +58,6 @@ const TOOLTIP_STYLE = {
   boxShadow: "0 4px 12px rgba(0, 0, 0, 0.3)",
   zIndex: 10000,
 } as const satisfies React.CSSProperties;
-const INFO_TOOLTIP_STYLE = {
-  padding: "6px 10px",
-  background: "#37373E",
-  color: "white",
-  fontSize: "12px",
-  fontWeight: 500,
-  borderRadius: "6px",
-  border: "1px solid #4C4C55",
-  boxShadow: "0 4px 12px rgba(0, 0, 0, 0.3)",
-} as const satisfies React.CSSProperties;
 
 function MasonryGridIcon() {
   return (
@@ -105,50 +95,6 @@ function SidebarCategoryHeader({
       {rightContent !== undefined && rightContent !== null && (
         <div className="tw-flex tw-items-center">{rightContent}</div>
       )}
-    </div>
-  );
-}
-
-function SidebarCategoryLabel({
-  label,
-  tooltipContent,
-}: {
-  readonly label: string;
-  readonly tooltipContent?: string | undefined;
-}) {
-  return (
-    <div className="tw-px-5 tw-pb-2 tw-pt-1 tw-text-[10px] tw-font-semibold tw-uppercase tw-leading-none tw-tracking-wide tw-text-iron-500">
-      <span className="tw-inline-flex tw-items-center tw-gap-x-1.5">
-        <span>{label}</span>
-        {tooltipContent && (
-          <span className="tw-relative tw-inline-flex tw-size-3 tw-items-center tw-justify-center">
-            <HoverCard
-              ariaLabel={tooltipContent}
-              placement="top"
-              delayShow={100}
-              delayHide={120}
-              hoverTransitionDelay={80}
-              offset={8}
-              openOnClick
-              closeOnContentClick
-              stopClickPropagation
-              triggerDisplay="inline-flex"
-              contentStyle={INFO_TOOLTIP_STYLE}
-              content={
-                <span className="tw-block tw-max-w-48">{tooltipContent}</span>
-              }
-            >
-              <button
-                type="button"
-                aria-label={tooltipContent}
-                className="tw-absolute tw-left-1/2 tw-top-1/2 tw-inline-flex tw-size-6 -tw-translate-x-1/2 -tw-translate-y-1/2 tw-items-center tw-justify-center tw-rounded-full tw-border-0 tw-bg-transparent tw-p-0 tw-text-iron-600 tw-transition-colors focus:tw-text-iron-300 focus:tw-outline-none active:tw-text-iron-300 desktop-hover:hover:tw-text-iron-400"
-              >
-                <FontAwesomeIcon icon={faInfoCircle} className="tw-size-3" />
-              </button>
-            </HoverCard>
-          </span>
-        )}
-      </span>
     </div>
   );
 }
@@ -657,6 +603,7 @@ const WebUnifiedWavesListWaves: React.FC<WebUnifiedWavesListWavesProps> = ({
                 <>
                   <SidebarCategoryLabel
                     label={t(SIDEBAR_LOCALE, "waves.sidebar.highlyRated")}
+                    paddingClassName="tw-px-5"
                     tooltipContent={highlyRatedInfoTooltip}
                   />
                   <HighlyRatedWavesToggle
@@ -669,6 +616,7 @@ const WebUnifiedWavesListWaves: React.FC<WebUnifiedWavesListWavesProps> = ({
                 !isCollapsed && (
                   <SidebarCategoryLabel
                     label={t(SIDEBAR_LOCALE, "waves.sidebar.highlyRated")}
+                    paddingClassName="tw-px-5"
                     tooltipContent={highlyRatedInfoTooltip}
                   />
                 )
